+++ tailwind.config.js
@@ -33,5 +33,5 @@ module.exports = {
       '2xl': '1536px',
     },
   },
-  plugins: [],
+  plugins: [require('@tailwindcss/typography')],
 }